Refactor remark to support current changes
