+++ src/main/java/seedu/address/logic/commands/RemarkCommand.java
@@ -10,7 +10,7 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.model.person.Person;
-import seedu.address.model.person.Remark;
+import seedu.address.model.patient.Remark;
 
 /**
  * Changes the remark of an existing person in the address book.
