+++ src/main/java/seedu/address/logic/parser/AddCommandParser.java
@@ -18,7 +18,7 @@ import seedu.address.model.patient.Name;
 import seedu.address.model.patient.Nric;
 import seedu.address.model.patient.Patient;
 import seedu.address.model.patient.Phone;
-import seedu.address.model.person.Remark;
+import seedu.address.model.patient.Remark;
 
 import seedu.address.model.tag.Tag;
