+++ src/main/java/seedu/address/logic/parser/RemarkCommandParser.java
@@ -8,7 +8,7 @@ import seedu.address.commons.core.index.Index;
 import seedu.address.commons.exceptions.IllegalValueException;
 import seedu.address.logic.commands.RemarkCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.person.Remark;
+import seedu.address.model.patient.Remark;
 
 public class RemarkCommandParser implements Parser<RemarkCommand> {
