+++ src/main/java/seedu/address/storage/JsonAdaptedPatient.java
@@ -15,7 +15,7 @@ import seedu.address.model.patient.Name;
 import seedu.address.model.patient.Nric;
 import seedu.address.model.patient.Patient;
 import seedu.address.model.patient.Phone;
-import seedu.address.model.person.Remark;
+import seedu.address.model.patient.Remark;
 import seedu.address.model.tag.Tag;
 
 /**
