+++ src/test/java/seedu/address/logic/commands/RemarkCommandTest.java
@@ -18,7 +18,7 @@ import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.UserPrefs;
 import seedu.address.model.person.Person;
-import seedu.address.model.person.Remark;
+import seedu.address.model.patient.Remark;
 import seedu.address.testutil.PersonBuilder;
 
 class RemarkCommandTest {
